Moved filtering and adjusting to "L1toL2"
