+++ .gitignore
@@ -10,6 +10,9 @@ dist/
 # Docs generation
 docs/_build/
 
+# local directory
+src/pypromice/local
+
 # Test output directories
 src/pypromice/test/TEST1/*
 src/pypromice/test/TEST2/*
